requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
gunicorn==23.0.0
sqlalchemy==2.0.45
asyncpg==0.31.0
alembic==1.17.2
pydantic==2.12.5
pydantic-settings==2.12.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.9
aiosmtplib==3.0.2
python-dotenv==1.0.1
redis==7.1.0
slowapi==0.1.9
loguru==0.7.3
python-json-logger==2.0.7
pillow==12.1.0